Added impl. for update of cluster memberships.

+++ src/main/scala/io/flatmap/ml/fuzzy/functions/package.scala
@@ -23,7 +23,7 @@ package object functions {
 
   def norm(x: DenseMatrix[Double]): Double = breeze.linalg.norm(x.toDenseVector)
 
-  def pow(x: DenseMatrix[Double], exp: Int): DenseMatrix[Double] =
+  def pow(x: DenseMatrix[Double], exp: Double): DenseMatrix[Double] =
     breeze.numerics.pow(x.toDenseVector, exp).toDenseMatrix.reshape(x.rows, x.cols)
 
 }
